refactor: moved loading to new module
